+++ Tests/Integration/DSCResources/Outlook2013.config.ps1
@@ -1,7 +1,11 @@
-Configuration WindowsOutlook2013_config
+Configuration Outlook2013_config
 {
     param
     (
+        [Parameter(Mandatory = $true)]
+        [string]
+        $Outlook2013,
+
         [Parameter(Mandatory = $true)]
         [version]
         $StigVersion
@@ -11,8 +15,9 @@ Configuration WindowsOutlook2013_config
 
     Node localhost
     {
-        WindowsOutlook2013 BaseLineSettings
+        Outlook2013 BaseLineSettings
         {
+            Outlook2013 = $Outlook2013
             StigVersion = $StigVersion
         }
     }