+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added the following STIGs
   * Microsoft Outlook 2013 STIG V1R12
+* Added the DSC Composite Resource for Windows Outlook2013 STIG
 
 ## 2.1.0.0